Rename src/ to sfta/
