+++ app.py
@@ -2,10 +2,10 @@ import logging
 
 import streamlit as st
 
-from src.context_manager import DeepRecurse
-from src.data_models.exceptions.base import FaultTreeTextException
-from src.data_models.fault_tree import FaultTree
-from src.output_handling import create_directory_robust, write_output_files
+from sfta.context_manager import DeepRecurse
+from sfta.data_models.exceptions.base import FaultTreeTextException
+from sfta.data_models.fault_tree import FaultTree
+from sfta.output_handling import create_directory_robust, write_output_files
 
 logging.basicConfig(level=logging.DEBUG)
