+++ sfta/data_models/exceptions/base.py
