+++ sfta/data_models/exceptions/event.py
