+++ sfta/data_models/exceptions/gate.py
