+++ sfta/data_models/nan.py
