+++ sfta/data_models/node.py
