+++ sfta/data_models/table.py
