+++ sfta/main.py
@@ -4,10 +4,10 @@ import argparse
 import os
 import sys
 
-from src.output_handling import create_directory_robust, write_output_files
-from src.context_manager import DeepRecurse
-from src.data_models.fault_tree import FaultTree
-from src.data_models.exceptions.base import FaultTreeTextException
+from sfta.output_handling import create_directory_robust, write_output_files
+from sfta.context_manager import DeepRecurse
+from sfta.data_models.fault_tree import FaultTree
+from sfta.data_models.exceptions.base import FaultTreeTextException
 
 
 __version__ = '0.6.2'
